Fix condition of test

+++ scripts/catalina-xslt.sh
@@ -236,14 +236,14 @@ if [ -n "$TOMCAT_HTTP_PORT" -o -n "$TOMCAT_HTTPS_PORT" -o -n "$TOMCAT_AJP_PORT" 
 fi
 
 echo "generating context-based XSL: $CONTEXT_XSL"
-if [ -n "$DEPLOY_CONTEXT" -n "$INPUT_CONTEXT" -n "$CONTEXT_XSL" -a -e "$INPUT_CONTEXT" -a ! -e "conf/Catalina/localhost/$DEPLOY_CONTEXT.xml" ]; then
+if [ -n "$DEPLOY_CONTEXT" -a -n "$INPUT_CONTEXT" -a -n "$CONTEXT_XSL" -a -e "$INPUT_CONTEXT" -a ! -e "conf/Catalina/localhost/$DEPLOY_CONTEXT.xml" ]; then
 	mkdir -p conf/Catalina/localhost
  	#XSL Processing
 	java -jar $JAR_PROC $CONTEXT_PARAMS $INPUT_CONTEXT $CONTEXT_XSL conf/Catalina/localhost/$DEPLOY_CONTEXT.xml
 fi
 
 echo "generating server-based XSL: $SERVER_XSL"
-if [ -n "$SERVER_XSL" -e "server-orig.xml" ]; then
+if [ -n "$SERVER_XSL" -a -e "server-orig.xml" ]; then
 	java -jar $JAR_PROC $SERVER_PARAMS server-orig.xml $SERVER_XSL conf/server.xml
  	#for faster next start time
  	mv "server-orig.xml" "server-orig.xml.bak"
